feat: update Book and BookContainer components to use new BookSearchResponse
