+++ src/mocks/mockBookList.ts
@@ -1,50 +1,3 @@
-type BookSearchResponse = {
-  title: string;
-  link: string;
-  author: string;
-  pubDate: string;
-  description: string;
-  isbn: string;
-  isbn13: string;
-  itemId: number;
-  priceSales: number;
-  priceStandard: number;
-  mallType: string;
-  stockStatus: string;
-  mileage: number;
-  cover: string;
-  categoryId: number;
-  categoryName: string;
-  publisher: string;
-  salesPoint: number;
-  adult: boolean;
-  fixedPrice: boolean;
-  customerReviewRank: number;
-  bestDuration?: string;
-  bestRank: number;
-  seriesInfo?: {
-    seriesId: number;
-    seriesLink: string;
-    seriesName: string;
-  };
-  subInfo: {};
-};
-
-type BookListResponse = {
-  version: string;
-  logo: string;
-  title: string;
-  link: string;
-  pubDate: string;
-  totalResults: number;
-  startIndex: number;
-  itemsPerPage: number;
-  query: string;
-  searchCategoryId: number;
-  searchCategoryName: string;
-  item: BookSearchResponse[];
-};
-
 const mockBookList = {
   version: '20131101',
   logo: 'http://image.aladin.co.kr/img/header/2011/aladin_logo_new.gif',
@@ -366,5 +319,4 @@ const mockBookList = {
   ],
 };
 
-export type { BookSearchResponse, BookListResponse };
 export { mockBookList };
