+++ src/types/aladin.type.ts
@@ -3,6 +3,53 @@
 https://docs.google.com/document/d/1mX-WxuoGs8Hy-QalhHcvuV17n50uGI2Sg_GHofgiePE/edit?tab=t.0#heading=h.wzyr9o7mof2u 
 */
 
+type BookSearchResponse = {
+  title: string;
+  link: string;
+  author: string;
+  pubDate: string;
+  description: string;
+  isbn: string;
+  isbn13: string;
+  itemId: number;
+  priceSales: number;
+  priceStandard: number;
+  mallType: string;
+  stockStatus: string;
+  mileage: number;
+  cover: string;
+  categoryId: number;
+  categoryName: string;
+  publisher: string;
+  salesPoint: number;
+  adult: boolean;
+  fixedPrice: boolean;
+  customerReviewRank: number;
+  bestDuration?: string;
+  bestRank: number;
+  seriesInfo?: {
+    seriesId: number;
+    seriesLink: string;
+    seriesName: string;
+  };
+  subInfo: {};
+};
+
+type BookListResponse = {
+  version: string;
+  logo: string;
+  title: string;
+  link: string;
+  pubDate: string;
+  totalResults: number;
+  startIndex: number;
+  itemsPerPage: number;
+  query: string;
+  searchCategoryId: number;
+  searchCategoryName: string;
+  item: BookSearchResponse[];
+};
+
 type BookLookUpResponse = {
   adult: boolean; //성인작품 여부
   author: string; //저자명
@@ -37,4 +84,4 @@ type BookLookUpResponse = {
   title: string; //상품명
 };
 
-export type { BookLookUpResponse };
+export type { BookSearchResponse, BookListResponse, BookLookUpResponse };
